// api/model/isFollowingResponse.ts
/**
 * twitter-api
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 2019-10-25T01:33:53Z
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface IsFollowingResponse { 
    /**
     * Boolean of the follow in question
     */
    isFollowing?: string;
}