// api/model/statusResponse.ts
/**
 * twitter-api
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 2019-11-01T00:49:12Z
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface StatusResponse { 
    /**
     * The status message
     */
    message?: string;
    /**
     * The status ownerHandle
     */
    ownerHandle?: string;
    /**
     * The status ownerHandle
     */
    profile?: string;
    /**
     * The status attachmentSrc
     */
    attachmentSrc?: string;
    /**
     * The status date
     */
    date?: string;
    /**
     * The status id
     */
    id?: string;
}