// api/model/updateProfileRequest.ts
/**
 * twitter-api
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 2019-11-19T00:53:19Z
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface UpdateProfileRequest { 
    /**
     * The profile user handle
     */
    handle?: string;
    /**
     * The profile src
     */
    src?: string;
}