// api/model/postStatusRequest.ts
/**
 * twitter-api
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 2019-10-30T17:33:34Z
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface PostStatusRequest { 
    /**
     * The status message
     */
    message?: string;
    /**
     * The status attachmentSrc
     */
    attachmentSrc?: string;
    /**
     * The status ownerHandle
     */
    ownerHandle?: string;
    profile?: string;
}